requirements.txt
discord.py
python-decouple
discord.py[voice]
